trying to run ym2203

+++ snd/burn_ym2203.h
@@ -10,7 +10,7 @@
 extern void BurnYM2203UpdateRequest();
 
 //int BurnYM2203Init(int num, int nClockFrequency, FM_IRQHANDLER IRQCallback, int (*StreamCallback)(int), double (*GetTimeCallback)(), int bAddSignal);
-INT32 BurnYM2203Init(INT32 num, INT16 *addr, INT32 nClockFrequency, FM_IRQHANDLER IRQCallback, INT32 (*StreamCallback)(INT32), double (*GetTimeCallback)(), INT32 bAddSignal);
+INT32 BurnYM2203Init(INT32 num, INT16 *addr, INT32 nClockFrequency, FM_IRQHANDLER IRQCallback, INT32 (*StreamCallback)(INT32), float (*GetTimeCallback)(), INT32 bAddSignal);
 
 void BurnYM2203SetRoute(int nChip, int nIndex, double nVolume, int nRouteDir);
 void BurnYM2203SetLeftVolume(int nChip, int nIndex, double nLeftVolume);
